Updated 0124

+++ 1102WinterPrerequire/BodyTrans.java
@@ -4,15 +4,15 @@ import java.util.Scanner;
 public class BodyTrans {
 	public static void main(String[] args) {
 		int cm, kg;
-		float inch, pound;
+		double inch, pound;
 
 		Scanner input = new Scanner(System.in);
 
 		cm = input.nextInt();
 		kg = input.nextInt();
 
-		inch = cm / 2.54f;
-		pound = kg / 0.454f;
+		inch = cm / 2.54;
+		pound = kg / 0.454;
 
 		System.out.printf("%.4f\n%.4f\n", inch, pound);
 	}
